double -> float32

+++ models/quantization_utils/quant_utils.py
@@ -216,28 +216,28 @@ class fixedpoint_mul(Function):
             ctx.z_scaling_factor = z_scaling_factor
 
             z_int = torch.round(pre_act / pre_act_scaling_factor)
-            _A = pre_act_scaling_factor.type(torch.double)
-            _B = (z_scaling_factor.type(torch.float)).type(torch.double)
+            _A = pre_act_scaling_factor.type(torch.float32)
+            _B = (z_scaling_factor.type(torch.float)).type(torch.float32)
             new_scale = _A / _B
             # print(new_scale)
             # exit()
             new_scale = reshape(new_scale)
 
             m, e = batch_frexp(new_scale)
-            output = z_int.type(torch.double) * m.type(torch.double)
+            output = z_int.type(torch.float32) * m.type(torch.float32)
             output = torch.round(output / (2.0 ** e))
 
             if identity is not None:
                 # needs addition of identity activation
                 wx_int = torch.round(identity / identity_scaling_factor)
 
-                _A = identity_scaling_factor.type(torch.double)
-                _B = (z_scaling_factor.type(torch.float)).type(torch.double)
+                _A = identity_scaling_factor.type(torch.float32)
+                _B = (z_scaling_factor.type(torch.float)).type(torch.float32)
                 new_scale = _A / _B
                 new_scale = reshape(new_scale)
 
                 m1, e1 = batch_frexp(new_scale)
-                output1 = wx_int.type(torch.double) * m1.type(torch.double)
+                output1 = wx_int.type(torch.float32) * m1.type(torch.float32)
                 output1 = torch.round(output1 / (2.0 ** e1))
 
                 output = output1 + output
